previous _reducer file deleted and a new file created in our current file structure setting the email into the redux store as a temporary means for the componentn to reach the redux store
